requirements.txt
Flask
psycopg2-binary
cohere
requests
beautifulsoup4
python-dotenv